# app/Http/Controllers/Section/SectionController.php
<?php
/**
 * Created by PhpStorm.
 * User: lucas
 * Date: 14/03/19
 * Time: 20:41
 */

namespace App\Http\Controllers\Section;


use App\Entities\Section\Section;
use App\FieldManagers\Section\SectionFieldManager;
use App\Http\Controllers\ApiController;
use App\Traits\ApiResponse;
use Illuminate\Http\Request;

class SectionController extends ApiController
{
    use ApiResponse;

    public function __construct(Section $model, SectionFieldManager $fieldManager)
    {
        $this->fieldManager = $fieldManager;
        $this->model = $model;
    }

    /**
     * @OA\Get(
     *     path="/sections",
     *     summary="Lista todas as seções",
     *     operationId="GetSections",
     *     tags={"sections"},
     *     @OA\Parameter(
     *         name="name",
     *         in="query",
     *         description="Nome da seção",
     *         required=false,
     *         @OA\Schema(
     *           type="string",
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="...",
     *     )
     * )
     */
    public function index(Request $request)
    {
        return parent::index($request); // TODO: Change the autogenerated stub
    }

    /**
     * @OA\Post(
     *     path="/sections",
     *     summary="Cria uma nova seção",
     *     operationId="StoreSection",
     *     tags={"sections"},
     *     @OA\Parameter(
     *         name="name",
     *         in="query",
     *         description="Nome da seção",
     *         required=true,
     *         @OA\Schema(
     *           type="string",
     *         )
     *     ),
     *     @OA\Parameter(
     *         name="description",
     *         in="query",
     *         description="Descrição da seção",
     *         required=true,
     *         @OA\Schema(
     *           type="string",
     *         )
     *     ),
     *    @OA\Parameter(
     *         name="order",
     *         in="query",
     *         description="Ordem da seção",
     *         required=false,
     *         @OA\Schema(
     *           type="integer",
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="...",
     *     )
     *
     * )
     */
    public function store(Request $request)
    {
        return parent::store($request); // TODO: Change the autogenerated stub
    }

    /**
     * @OA\Get(
     *     path="/sections/{section_id}",
     *     summary="Lista uma seções",
     *     operationId="GetSection",
     *     tags={"sections"},
     *     @OA\Parameter(
     *         name="section_id",
     *         in="path",
     *         description="Id da seção",
     *         required=true,
     *         @OA\Schema(
     *           type="integer"
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="...",
     *     )
     * )
     */
    public function show(int $section_id)
    {
        return parent::show($section_id); // TODO: Change the autogenerated stub
    }

    /**
     * @OA\Put(
     *     path="/sections/{section_id}",
     *     summary="Atualiza uma seção",
     *     operationId="UpdateSection",
     *     tags={"sections"},
     *     @OA\Parameter(
     *         name="section_id",
     *         in="path",
     *         description="Id da seção",
     *         required=true,
     *         @OA\Schema(
     *           type="integer",
     *         )
     *     ),
     *     @OA\Parameter(
     *         name="name",
     *         in="query",
     *         description="Nome da seção",
     *         required=false,
     *         @OA\Schema(
     *           type="string",
     *         )
     *     ),
     *     @OA\Parameter(
     *         name="description",
     *         in="query",
     *         description="Descrição da seção",
     *         required=false,
     *         @OA\Schema(
     *           type="string",
     *         )
     *     ),
     *    @OA\Parameter(
     *         name="order",
     *         in="query",
     *         description="Ordem da seção",
     *         required=false,
     *         @OA\Schema(
     *           type="integer",
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="...",
     *     )
     *
     * )
     */
    public function update(Request $request, int $section_id)
    {
        return parent::update($request, $section_id); // TODO: Change the autogenerated stub
    }

    /**
     * @OA\Delete(
     *     path="/sections/{section_id}",
     *     summary="Apaga uma seção",
     *     operationId="DeleteSection",
     *     tags={"sections"},
     *     @OA\Parameter(
     *         name="section_id",
     *         in="path",
     *         description="Id da seção",
     *         required=true,
     *         @OA\Schema(
     *           type="number"
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="...",
     *     )
     * )
     */
    public function destroy(int $id)
    {
        return parent::destroy($id); // TODO: Change the autogenerated stub
    }
}