# public/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: 'My first API'
  contact:
    email: support@example.com
  version: 1.0.0
paths:
...
  '/sections/{section_id}/lessons':
    get:
      tags:
        - lessons
      summary: 'Lista todas as lições'
      operationId: GetLessons
      parameters:
        -
          name: section_id
          in: path
          description: 'Id da seção'
          required: true
          schema:
            type: integer
        -
          name: includes
          in: query
          description: 'Faz o include das relações'
          required: false
          schema:
            type: array
            items:
              type: string
        -
          name: name
          in: query
          description: 'Nome da lição'
          required: false
          schema:
            type: string
      responses:
        '200':
          description: ...
    post:
      tags:
        - lessons
      summary: 'Cria uma nova lição'
      operationId: StoreLesson
      parameters:
        -
          name: section_id
          in: path
          description: 'Id da seção no path'
          required: true
          schema:
            type: integer
        -
          name: name
          in: query
          description: 'Nome da lição'
          required: true
          schema:
            type: string
        -
          name: content
          in: query
          description: 'Conteúdo da lição'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ...
  '/sections/{section_id}/lessons/{lesson_id}':
    get:
      tags:
        - lessons
      summary: 'Lista uma lição'
      operationId: GetLesson
      parameters:
        -
          name: section_id
          in: path
          description: 'Id da seção'
          required: true
          schema:
            type: integer
        -
          name: lesson_id
          in: path
          description: 'Id da lição'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ...
    put:
      tags:
        - lessons
      summary: 'Atualiza uma lição'
      operationId: UpdateLesson
      parameters:
        -
          name: section_id
          in: path
          description: 'Id da seção'
          required: true
          schema:
            type: integer
        -
          name: lesson_id
          in: path
          description: 'Id da lição'
          required: true
          schema:
            type: integer
        -
          name: name
          in: query
          description: 'Nome da lição'
          required: true
          schema:
            type: string
        -
          name: content
          in: query
          description: 'Conteúdo da lição'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ...
    delete:
      tags:
        - lessons
      summary: 'Lista uma lição'
      operationId: DeleteLesson
      parameters:
        -
          name: section_id
          in: path
          description: 'Id da seção'
          required: true
          schema:
            type: integer
        -
          name: lesson_id
          in: path
          description: 'Id da lição'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ...
  /sections:
    get:
      tags:
        - sections
      summary: 'Lista todas as seções'
      operationId: GetSections
      parameters:
        -
          name: name
          in: query
          description: 'Nome da seção'
          required: false
          schema:
            type: string
      responses:
        '200':
          description: ...
    post:
      tags:
        - sections
      summary: 'Cria uma nova seção'
      operationId: StoreSection
      parameters:
        -
          name: name
          in: query
          description: 'Nome da seção'
          required: true
          schema:
            type: string
        -
          name: description
          in: query
          description: 'Descrição da seção'
          required: true
          schema:
            type: string
        -
          name: order
          in: query
          description: 'Ordem da seção'
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: ...
  '/sections/{section_id}':
    get:
      tags:
        - sections
      summary: 'Lista uma seções'
      operationId: GetSection
      parameters:
        -
          name: section_id
          in: path
          description: 'Id da seção'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ...
    put:
      tags:
        - sections
      summary: 'Atualiza uma seção'
      operationId: UpdateSection
      parameters:
        -
          name: section_id
          in: path
          description: 'Id da seção'
          required: true
          schema:
            type: integer
        -
          name: name
          in: query
          description: 'Nome da seção'
          required: false
          schema:
            type: string
        -
          name: description
          in: query
          description: 'Descrição da seção'
          required: false
          schema:
            type: string
        -
          name: order
          in: query
          description: 'Ordem da seção'
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: ...
    delete:
      tags:
        - sections
      summary: 'Apaga uma seção'
      operationId: DeleteSection
      parameters:
        -
          name: section_id
          in: path
          description: 'Id da seção'
          required: true
          schema:
            type: number
      responses:
        '200':
          description: ...
...
components:
  securitySchemes:
    apiToken:
      type: apiKey
      description: 'Token do client para consumo da API'
      name: Authorization
      in: header
      scheme: bearer